move event template actions to api & update event templates when they are changed by another user
