+++ client/actions/events/api.js
@@ -1430,6 +1430,37 @@ const receiveCalendars = (calendars) => ({
     payload: calendars,
 });
 
+const fetchEventTemplates = () => (dispatch, getState, {api}) => {
+    api('recent_events_template').query()
+        .then((res) => {
+            dispatch({type: 'RECEIVE_EVENT_TEMPLATES', payload: res._items});
+        });
+};
+
+const updateEventTemplate = (template, updates) => (dispatch, getState, {api}) => {
+    api.update('events_template', template, updates)
+        .then(() => {
+            dispatch(fetchEventTemplates());
+        });
+};
+
+const removeEventTemplate = (template) => (dispatch, getState, {api}) => {
+    api.remove(template, {}, 'events_template')
+        .then(() => {
+            dispatch(fetchEventTemplates());
+        });
+};
+
+const saveEventTemplate = (tamplateName, eventId) => (dispatch, getState, {api}) => {
+    api('events_template').save({
+        template_name: tamplateName,
+        based_on_event: eventId,
+    })
+        .then(() => {
+            dispatch(fetchEventTemplates());
+        });
+};
+
 // eslint-disable-next-line consistent-this
 const self = {
     loadEventsByRecurrenceId,
@@ -1467,6 +1498,10 @@ const self = {
     receiveCalendars,
     fetchEventFiles,
     removeFile,
+    fetchEventTemplates,
+    saveEventTemplate,
+    updateEventTemplate,
+    removeEventTemplate,
 };
 
 export default self;
